test(txc-parser): refactor TXC parser tests into separate files to ease readability
These test files were over 500 lines long making them hard to read and add to

Splitting them out by section will increase maintainability

- Refactor tests into multiple files for:
  - VehicleJourneys
  - RouteSections
  - JourneyPatternSection
