+++ tests/timetables_etl/etl/txc/parser/txc_vehicle_journey/operating_profile/test_txc_operating_profile_regular_days.py
