+++ tests/timetables_etl/etl/txc/parser/txc_vehicle_journey/test_txc_vehicle_journeys_flexible.py
@@ -15,7 +15,7 @@ from common_layer.txc.parser.vehicle_journeys_flexible import (
 from lxml import etree
 from pydantic import ValidationError
 
-from .utils import assert_model_equal
+from ..utils import assert_model_equal
 
 
 @pytest.mark.parametrize(
